Update tutorial 5 and code demo
> updated tutorial steps
> fixed code issues
> updated screenshots
> updated code solution
