+++ demos/03-add-msgraph/GraphTutorial/app/build.gradle
@@ -3,7 +3,7 @@ apply plugin: 'com.android.application'
 android {
     compileSdkVersion 28
     defaultConfig {
-        applicationId "com.example.graphtutorial"
+        applicationId "com.microsoft.graphtutorial"
         minSdkVersion 27
         targetSdkVersion 28
         versionCode 1
@@ -22,13 +22,14 @@ android {
 }
 
 dependencies {
+    implementation 'com.android.support:design:28.0.0'
+    implementation 'com.microsoft.identity.client:msal:0.2.2'
+    implementation 'com.microsoft.graph:microsoft-graph:1.4.0'
+
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation 'com.android.support:appcompat-v7:28.0.0'
-    implementation 'com.android.support:design:28.0.0'
     implementation 'com.android.support.constraint:constraint-layout:1.1.3'
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
-    implementation 'com.microsoft.graph:microsoft-graph:1.1.+'
-    implementation 'com.microsoft.identity.client:msal:0.2.+'
 }
