+++ demos/03-add-msgraph/GraphTutorial/app/src/androidTest/java/com/microsoft/graphtutorial/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.graphtutorial;
+package com.microsoft.graphtutorial;
 
 import android.content.Context;
 import android.support.test.InstrumentationRegistry;
@@ -16,11 +16,11 @@ import static org.junit.Assert.*;
  */
 @RunWith(AndroidJUnit4.class)
 public class ExampleInstrumentedTest {
-    @Test
-    public void useAppContext() {
-        // Context of the app under test.
-        Context appContext = InstrumentationRegistry.getTargetContext();
+  @Test
+  public void useAppContext() {
+    // Context of the app under test.
+    Context appContext = InstrumentationRegistry.getTargetContext();
 
-        assertEquals("com.example.graphtutorial", appContext.getPackageName());
-    }
+    assertEquals("com.microsoft.graphtutorial", appContext.getPackageName());
+  }
 }
