--- demos/03-add-msgraph/GraphTutorial/app/src/main/java/com/example/graphtutorial/AuthenticationHelper.java
@@ -1,64 +1,0 @@
-package com.example.graphtutorial;
-
-import android.app.Activity;
-import android.content.Context;
-import android.content.Intent;
-
-import com.microsoft.identity.client.AuthenticationCallback;
-import com.microsoft.identity.client.IAccount;
-import com.microsoft.identity.client.PublicClientApplication;
-
-// Singleton class - the app only needs a single instance
-// of PublicClientApplication
-public class AuthenticationHelper {
-    private static AuthenticationHelper INSTANCE = null;
-    private PublicClientApplication mPCA = null;
-    private String[] mScopes;
-
-    private AuthenticationHelper(Context ctx) {
-        String appId = ctx.getResources().getString(R.string.oauth_app_id);
-        mScopes = ctx.getResources().getStringArray(R.array.oauth_scopes);
-        mPCA = new PublicClientApplication(ctx, appId);
-    }
-
-    public static synchronized AuthenticationHelper getInstance(Context ctx) {
-        if (INSTANCE == null) {
-            INSTANCE = new AuthenticationHelper(ctx);
-        }
-
-        return INSTANCE;
-    }
-
-    // Version called from fragments. Does not create an
-    // instance if one doesn't exist
-    public static synchronized AuthenticationHelper getInstance() {
-        if (INSTANCE == null) {
-            throw new IllegalStateException(
-                    "AuthenticationHelper has not been initialized from MainActivity");
-        }
-
-        return INSTANCE;
-    }
-
-    public boolean hasAccount() {
-        return !mPCA.getAccounts().isEmpty();
-    }
-
-    public void handleRedirect(int requestCode, int resultCode, Intent data) {
-        mPCA.handleInteractiveRequestRedirect(requestCode, resultCode, data);
-    }
-
-    public void acquireTokenInteractively(Activity activity, AuthenticationCallback callback) {
-        mPCA.acquireToken(activity, mScopes, callback);
-    }
-
-    public void acquireTokenSilently(AuthenticationCallback callback) {
-        mPCA.acquireTokenSilentAsync(mScopes, mPCA.getAccounts().get(0), callback);
-    }
-
-    public void signOut() {
-        for (IAccount account : mPCA.getAccounts()) {
-            mPCA.removeAccount(account);
-        }
-    }
-}
