+++ demos/03-add-msgraph/GraphTutorial/app/src/main/java/com/microsoft/graphtutorial/CalendarFragment.java
@@ -1,0 +1,119 @@
+package com.microsoft.graphtutorial;
+
+import android.os.Bundle;
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
+import android.support.v4.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import android.util.Log;
+import android.widget.ListView;
+import android.widget.ProgressBar;
+import com.microsoft.graph.concurrency.ICallback;
+import com.microsoft.graph.core.ClientException;
+import com.microsoft.graph.models.extensions.Event;
+import com.microsoft.graph.requests.extensions.IEventCollectionPage;
+import com.microsoft.identity.client.AuthenticationCallback;
+import com.microsoft.identity.client.AuthenticationResult;
+import com.microsoft.identity.client.exception.MsalException;
+import java.util.List;
+
+public class CalendarFragment extends Fragment {
+  private List<Event> mEventList = null;
+  private ProgressBar mProgress = null;
+
+  @Nullable
+  @Override
+  public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+    return inflater.inflate(R.layout.fragment_calendar, container, false);
+  }
+
+  private void showProgressBar() {
+    getActivity().runOnUiThread(new Runnable() {
+      @Override
+      public void run() {
+        mProgress.setVisibility(View.VISIBLE);
+      }
+    });
+  }
+
+  private void hideProgressBar() {
+    getActivity().runOnUiThread(new Runnable() {
+      @Override
+      public void run() {
+        mProgress.setVisibility(View.GONE);
+      }
+    });
+  }
+
+  private ICallback<IEventCollectionPage> getEventsCallback() {
+    return new ICallback<IEventCollectionPage>() {
+      @Override
+      public void success(IEventCollectionPage iEventCollectionPage) {
+        mEventList = iEventCollectionPage.getCurrentPage();
+
+        addEventsToList();
+
+        // Temporary for debugging
+        String jsonEvents = GraphHelper.getInstance().serializeObject(mEventList);
+        Log.d("GRAPH", jsonEvents);
+
+        hideProgressBar();
+      }
+
+      @Override
+      public void failure(ClientException ex) {
+        Log.e("GRAPH", "Error getting events", ex);
+        hideProgressBar();
+      }
+    };
+  }
+
+  @Override
+  public void onCreate(@Nullable Bundle savedInstanceState) {
+    super.onCreate(savedInstanceState);
+
+    mProgress = getActivity().findViewById(R.id.progressbar);
+    showProgressBar();
+
+    // Get a current access token
+    AuthenticationHelper.getInstance()
+            .acquireTokenSilently(new AuthenticationCallback() {
+              @Override
+              public void onSuccess(AuthenticationResult authenticationResult) {
+                final GraphHelper graphHelper = GraphHelper.getInstance();
+
+                // Get the user's events
+                graphHelper.getEvents(authenticationResult.getAccessToken(),
+                        getEventsCallback());
+              }
+
+              @Override
+              public void onError(MsalException exception) {
+                Log.e("AUTH", "Could not get token silently", exception);
+                hideProgressBar();
+              }
+
+              @Override
+              public void onCancel() {
+                hideProgressBar();
+              }
+            });
+  }
+
+  private void addEventsToList() {
+    getActivity().runOnUiThread(new Runnable() {
+      @Override
+      public void run() {
+        ListView eventListView = getView().findViewById(R.id.eventlist);
+
+        EventListAdapter listAdapter = new EventListAdapter(getActivity(),
+                R.layout.event_list_item, mEventList);
+
+        eventListView.setAdapter(listAdapter);
+      }
+    });
+  }
+}
