+++ demos/03-add-msgraph/GraphTutorial/app/src/main/java/com/microsoft/graphtutorial/EventListAdapter.java
@@ -1,0 +1,84 @@
+package com.microsoft.graphtutorial;
+
+import android.content.Context;
+import android.support.annotation.NonNull;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.ArrayAdapter;
+import android.widget.TextView;
+
+import com.microsoft.graph.models.extensions.DateTimeTimeZone;
+import com.microsoft.graph.models.extensions.Event;
+
+import java.time.LocalDateTime;
+import java.time.ZoneId;
+import java.time.ZonedDateTime;
+import java.time.format.DateTimeFormatter;
+import java.time.format.FormatStyle;
+import java.util.List;
+import java.util.TimeZone;
+
+public class EventListAdapter extends ArrayAdapter<Event> {
+  private Context mContext;
+  private int mResource;
+  private ZoneId mLocalTimeZoneId;
+
+  // Used for the ViewHolder pattern
+  // https://developer.android.com/training/improving-layouts/smooth-scrolling
+  static class ViewHolder {
+    TextView subject;
+    TextView organizer;
+    TextView start;
+    TextView end;
+  }
+
+  public EventListAdapter(Context context, int resource, List<Event> events) {
+    super(context, resource, events);
+    mContext = context;
+    mResource = resource;
+    mLocalTimeZoneId = TimeZone.getDefault().toZoneId();
+  }
+
+  @NonNull
+  @Override
+  public View getView(int position, View convertView, ViewGroup parent) {
+    Event event = getItem(position);
+
+    ViewHolder holder;
+
+    if (convertView == null) {
+      LayoutInflater inflater = LayoutInflater.from(mContext);
+      convertView = inflater.inflate(mResource, parent, false);
+
+      holder = new ViewHolder();
+      holder.subject = convertView.findViewById(R.id.eventsubject);
+      holder.organizer = convertView.findViewById(R.id.eventorganizer);
+      holder.start = convertView.findViewById(R.id.eventstart);
+      holder.end = convertView.findViewById(R.id.eventend);
+
+      convertView.setTag(holder);
+    } else {
+      holder = (ViewHolder) convertView.getTag();
+    }
+
+    holder.subject.setText(event.subject);
+    holder.organizer.setText(event.organizer.emailAddress.name);
+    holder.start.setText(getLocalDateTimeString(event.start));
+    holder.end.setText(getLocalDateTimeString(event.end));
+
+    return convertView;
+  }
+
+  // Convert Graph's DateTimeTimeZone format to
+  // a LocalDateTime, then return a formatted string
+  private String getLocalDateTimeString(DateTimeTimeZone dateTime) {
+    ZonedDateTime localDateTime = LocalDateTime.parse(dateTime.dateTime)
+            .atZone(ZoneId.of(dateTime.timeZone))
+            .withZoneSameInstant(mLocalTimeZoneId);
+
+    return String.format("%s %s",
+            localDateTime.format(DateTimeFormatter.ofLocalizedDate(FormatStyle.MEDIUM)),
+            localDateTime.format(DateTimeFormatter.ofLocalizedTime(FormatStyle.SHORT)));
+  }
+}
