+++ demos/03-add-msgraph/GraphTutorial/app/src/main/res/layout/fragment_calendar.xml
@@ -1,13 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
+  android:layout_width="match_parent"
+  android:layout_height="match_parent">
 
-    <ListView
-        android:id="@+id/eventlist"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:divider="@color/colorPrimary"
-        android:dividerHeight="1dp" />
+  <ListView
+    android:id="@+id/eventlist"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:divider="@color/colorPrimary"
+    android:dividerHeight="1dp" />
 
 </RelativeLayout>
