+++ demos/03-add-msgraph/GraphTutorial/app/src/test/java/com/microsoft/graphtutorial/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.graphtutorial;
+package com.microsoft.graphtutorial;
 
 import org.junit.Test;
 
@@ -10,8 +10,8 @@ import static org.junit.Assert.*;
  * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
  */
 public class ExampleUnitTest {
-    @Test
-    public void addition_isCorrect() {
-        assertEquals(4, 2 + 2);
-    }
+  @Test
+  public void addition_isCorrect() {
+    assertEquals(4, 2 + 2);
+  }
 }
